Vector2D operators Multiple colliders on entity Relative position on Collider2D

+++ include/Utils.hpp
@@ -6,7 +6,7 @@ namespace utils {
 
 	inline float HireTimeInSeconds() 
 	{
-		float t = SDL_GetTicks64();
+		float t = (float)SDL_GetTicks64();
 		t *= 0.001f;
 		return t;
 	}
